更新

+++ foreign/src/main/java/com/github/teruteru128/foreign/prime/search/Result.java
@@ -1,5 +1,5 @@
 package com.github.teruteru128.foreign.prime.search;
 
-public record Result(long step, int result) {
+public record Result(int step, int result) {
 
 }
